♻️ [refactor] 랜딩페이지 컴포넌트 분리
랜딩페이지 컴포넌트 분리

fixes: 추가 수정 가능성 있음
