+++ src/components/onBoarding/OnBoarding.module.css
@@ -7,7 +7,7 @@
 }
 
 .joinContentWrapper {
-  @apply absolute top-[40%] left-1/2 -translate-x-1/2 -translate-y-1/2 flex flex-col gap-3 text-center;
+  @apply absolute top-[40%] left-1/2 -translate-x-1/2 -translate-y-1/2 w-full flex flex-col gap-3 text-center;
 }
 
 .joinContent {
@@ -19,9 +19,9 @@
 }
 
 .joinButton {
-  @apply relative left-1/2 -translate-x-1/2 top-5 bg-primary w-[100%] py-5 rounded;
+  @apply relative left-1/2 -translate-x-1/2 top-5 bg-primary w-[30%] py-5 rounded;
   @apply text-white text-2xl font-semibold;
-  @apply before:bg-[url('/assets/tving-symbol.svg')] before:bg-no-repeat before:absolute before:top-0 before:left-0 before:w-20 before:h-full;
+  @apply before:bg-[url('/assets/tving-symbol.svg')] before:bg-no-repeat before:bg-contain before:absolute before:top-0 before:left-0 before:w-20 before:h-full max-md:w-[50%] max-lg:w-[40%];
 }
 
 .onlyTaing {
